[intel-npu] Limiting getDevice function to existing devices (#30368)
### Details:
- currently npu plugin will return the first existing npu device in the
device list all the time, regardless to what parameter is coming from
app level. This can cause the illusion that we have multiple devices,
i.e. user will get valid data for NPU.0, NPU.1, NPU.2, NPU.RANDOMSTRING
creating the confustion that there a multiple devices
- this PR implements the logic in getDevice function to check and pick
the requested device. If the requested device is not found, it throws a
corresponding error
- It can take multiple valid forms of input and search for them and
throw corresponding error if not found:
         - Index in the NPU device array (if exits): NPU.0, NPU.1 etc
         - Arch number: NPU.3720, NPU.4000
         - Custom name: NPU.RANDOMSTRING 
- There is 1 caveat: in order to be able to differenciate between arch
number and index, it makes the assumption that up to 3 digit numbers are
supposed to be indexes, while more than 3 digits are supposed to be arch
numbers

+ piggybacking a small addition to npu-documentation

### Tickets:
 - *[ticket-id](https://jira.devtools.intel.com/browse/CVS-165833)*
